chore: optimize
